# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Express Demo",
        "description": "Express Demo",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "/",
            "description": "Cloud server"
        }
       
    ],
    "tags": [
        {
            "name": "Users",
            "description": "API for users"
        }
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/v1/getUserList": {
            "get": {
                "tags": [
                    "User List"
                ],
                "summary": "Get all user list",
                "responses": {
                    "200": {
                        "description": "OK"
                        
                    }
                }
            }
            
        },
        "/api/v1/registerUser": {
            
            "post": {
                "tags": [
                    "registerUser"
                ],
                "summary": "Create a new user in system",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/users"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        },
        "/api/v1/loginUser": {
            
            "post": {
                "tags": [
                    "loginUser"
                ],
                "summary": "User login",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/users_login"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed. Bad post data."
                    }
                }
            }
        }
        
    },
    "definitions": {
        
        "users": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "users_login": {
            "type": "object",
            "properties": {
                
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
        
    }
}